The package analyzer now consumes the declaration environment directly—imported functions are no longer represented by synthetic AST stubs. The end-to-end regression now compiles an imported overloaded generic call all the way through code generation, and mutually recursive modules are checked against the precollected graph.

+++ compiler/frontend/hir/src/lib.rs
@@ -16,7 +16,7 @@ pub struct HirId(pub u32);
 pub struct BindingId(pub u32);
 
 #[derive(Debug, Clone, Copy, PartialEq, Eq, PartialOrd, Ord, Hash)]
-pub struct FunctionId(pub u32);
+pub struct FunctionId(pub u128);
 
 #[derive(Debug, Clone, Copy, PartialEq, Eq, PartialOrd, Ord, Hash)]
 pub struct DeclaredTypeId(pub u32);
